refac: seperate concerns into different files
