+++ client/src/game/constants.ts
@@ -1,0 +1,11 @@
+import { Piece } from "../types/types";
+
+export const ROWS = 10;
+export const COLS = 8;
+
+export const PIECES: Piece[] = [
+  { id: 1, width: 1, element: "stone1" },
+  { id: 2, width: 2, element: "stone2" },
+  { id: 3, width: 3, element: "stone3" },
+  { id: 4, width: 4, element: "stone4" },
+];
